♻️ Deler tester på ulike filer for å gjøre det mer lesbart
