+++ src/test/kotlin/no/nav/tilleggsstonader/sak/vedtak/boutgifter/BoutgifterTestUtil.kt
@@ -11,165 +11,37 @@ import no.nav.tilleggsstonader.sak.vedtak.domain.GeneriskVedtak
 import no.nav.tilleggsstonader.sak.vedtak.domain.InnvilgelseBoutgifter
 import no.nav.tilleggsstonader.sak.vedtak.domain.TypeBoutgift
 import no.nav.tilleggsstonader.sak.vedtak.domain.Vedtaksperiode
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.VilkårperiodeTestUtil.aktivitet
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.VilkårperiodeTestUtil.målgruppe
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.AktivitetType
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.Vilkårperioder
 import java.time.LocalDate
 import java.util.UUID
 
 object BoutgifterTestUtil {
     val vedtaksperiodeIdFørstegangsbehandling = UUID.randomUUID()
 
-    val utgiftMidlertidigOvernatting: Map<TypeBoutgift, List<UtgiftBeregningBoutgifter>> =
-        mapOf(
-            TypeBoutgift.UTGIFTER_OVERNATTING to
+    val vilkårperioder =
+        Vilkårperioder(
+            målgrupper =
                 listOf(
-                    UtgiftBeregningBoutgifter(
+                    målgruppe(
                         fom = LocalDate.of(2025, 1, 1),
-                        tom = LocalDate.of(2025, 1, 31),
-                        utgift = 3000,
+                        tom = LocalDate.of(2025, 10, 31),
                     ),
                 ),
-        )
-
-    val løpendeUtgifterEnBolig: Map<TypeBoutgift, List<UtgiftBeregningBoutgifter>> =
-        mapOf(
-            TypeBoutgift.LØPENDE_UTGIFTER_EN_BOLIG to
+            aktiviteter =
                 listOf(
-                    UtgiftBeregningBoutgifter(
+                    aktivitet(
                         fom = LocalDate.of(2025, 1, 1),
-                        tom = LocalDate.of(2025, 3, 31),
-                        utgift = 3000,
+                        tom = LocalDate.of(2025, 10, 31),
                     ),
                 ),
-        )
-
-    val løpendeUtgifterToBoliger: Map<TypeBoutgift, List<UtgiftBeregningBoutgifter>> =
-        mapOf(
-            TypeBoutgift.LØPENDE_UTGIFTER_TO_BOLIGER to
-                listOf(
-                    UtgiftBeregningBoutgifter(
-                        fom = LocalDate.of(2025, 1, 1),
-                        tom = LocalDate.of(2025, 3, 31),
-                        utgift = 3000,
-                    ),
-                ),
-        )
-
-    val beregningsresultatFørstegangsbehandlingMidlertidigOvernatting =
-        listOf(
-            BeregningsresultatForLøpendeMåned(
-                grunnlag =
-                    Beregningsgrunnlag(
-                        fom = LocalDate.of(2025, 1, 1),
-                        tom = LocalDate.of(2025, 1, 31),
-                        utbetalingsdato = LocalDate.of(2025, 1, 1),
-                        utgifter = utgiftMidlertidigOvernatting,
-                        makssats = 4953,
-                        makssatsBekreftet = true,
-                        målgruppe = FaktiskMålgruppe.NEDSATT_ARBEIDSEVNE,
-                        aktivitet = AktivitetType.TILTAK,
-                    ),
-                delAvTidligereUtbetaling = false,
-            ),
-        )
-
-    val beregningsresultatFørstegangsbehandlingLøpendeUtgifterEnBolig =
-        listOf(
-            BeregningsresultatForLøpendeMåned(
-                grunnlag =
-                    Beregningsgrunnlag(
-                        fom = LocalDate.of(2025, 1, 1),
-                        tom = LocalDate.of(2025, 1, 31),
-                        utbetalingsdato = LocalDate.of(2025, 1, 1),
-                        utgifter = løpendeUtgifterEnBolig,
-                        makssats = 4953,
-                        makssatsBekreftet = true,
-                        målgruppe = FaktiskMålgruppe.NEDSATT_ARBEIDSEVNE,
-                        aktivitet = AktivitetType.TILTAK,
-                    ),
-                delAvTidligereUtbetaling = false,
-            ),
-            BeregningsresultatForLøpendeMåned(
-                grunnlag =
-                    Beregningsgrunnlag(
-                        fom = LocalDate.of(2025, 2, 1),
-                        tom = LocalDate.of(2025, 2, 28),
-                        utbetalingsdato = LocalDate.of(2025, 2, 1),
-                        utgifter = løpendeUtgifterEnBolig,
-                        makssats = 4953,
-                        makssatsBekreftet = true,
-                        målgruppe = FaktiskMålgruppe.NEDSATT_ARBEIDSEVNE,
-                        aktivitet = AktivitetType.TILTAK,
-                    ),
-                delAvTidligereUtbetaling = false,
-            ),
-            BeregningsresultatForLøpendeMåned(
-                grunnlag =
-                    Beregningsgrunnlag(
-                        fom = LocalDate.of(2025, 3, 1),
-                        tom = LocalDate.of(2025, 3, 31),
-                        utbetalingsdato = LocalDate.of(2025, 3, 1),
-                        utgifter = løpendeUtgifterEnBolig,
-                        makssats = 4953,
-                        makssatsBekreftet = true,
-                        målgruppe = FaktiskMålgruppe.NEDSATT_ARBEIDSEVNE,
-                        aktivitet = AktivitetType.TILTAK,
-                    ),
-                delAvTidligereUtbetaling = false,
-            ),
-        )
-
-    val beregningsresultatFørstegangsbehandlingLøpendeUtgifterToBoliger =
-        listOf(
-            BeregningsresultatForLøpendeMåned(
-                grunnlag =
-                    Beregningsgrunnlag(
-                        fom = LocalDate.of(2025, 1, 1),
-                        tom = LocalDate.of(2025, 1, 31),
-                        utbetalingsdato = LocalDate.of(2025, 1, 1),
-                        utgifter = løpendeUtgifterToBoliger,
-                        makssats = 4953,
-                        makssatsBekreftet = true,
-                        målgruppe = FaktiskMålgruppe.NEDSATT_ARBEIDSEVNE,
-                        aktivitet = AktivitetType.TILTAK,
-                    ),
-                delAvTidligereUtbetaling = false,
-            ),
-            BeregningsresultatForLøpendeMåned(
-                grunnlag =
-                    Beregningsgrunnlag(
-                        fom = LocalDate.of(2025, 2, 1),
-                        tom = LocalDate.of(2025, 2, 28),
-                        utbetalingsdato = LocalDate.of(2025, 2, 1),
-                        utgifter = løpendeUtgifterToBoliger,
-                        makssats = 4953,
-                        makssatsBekreftet = true,
-                        målgruppe = FaktiskMålgruppe.NEDSATT_ARBEIDSEVNE,
-                        aktivitet = AktivitetType.TILTAK,
-                    ),
-                delAvTidligereUtbetaling = false,
-            ),
-            BeregningsresultatForLøpendeMåned(
-                grunnlag =
-                    Beregningsgrunnlag(
-                        fom = LocalDate.of(2025, 3, 1),
-                        tom = LocalDate.of(2025, 3, 31),
-                        utbetalingsdato = LocalDate.of(2025, 3, 1),
-                        utgifter = løpendeUtgifterToBoliger,
-                        makssats = 4953,
-                        makssatsBekreftet = true,
-                        målgruppe = FaktiskMålgruppe.NEDSATT_ARBEIDSEVNE,
-                        aktivitet = AktivitetType.TILTAK,
-                    ),
-                delAvTidligereUtbetaling = false,
-            ),
         )
 
     fun innvilgelseBoutgifter(
         vedtaksperioder: List<Vedtaksperiode>,
-        beregningsresultat: BeregningsresultatBoutgifter =
-            BeregningsresultatBoutgifter(
-                beregningsresultatFørstegangsbehandlingMidlertidigOvernatting,
-            ),
+        beregningsresultat: BeregningsresultatBoutgifter,
     ) = GeneriskVedtak(
         behandlingId = BehandlingId.random(),
         data =
